backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-generativeai==0.8.3
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1